Moved views into pro_bono.py
