+++ corehq/apps/domain/views/pro_bono.py
@@ -133,3 +133,79 @@ from corehq.apps.hqwebapp.signals import clear_login_attempts
 from corehq.apps.ota.models import MobileRecoveryMeasure
 import six
 from six.moves import map
+
+
+class ProBonoMixin(object):
+    page_title = ugettext_lazy("Pro-Bono Application")
+    is_submitted = False
+
+    url_name = None
+
+    @property
+    def requesting_domain(self):
+        raise NotImplementedError
+
+    @property
+    @memoized
+    def pro_bono_form(self):
+        if self.request.method == 'POST':
+            return ProBonoForm(self.use_domain_field, self.request.POST)
+        return ProBonoForm(self.use_domain_field)
+
+    @property
+    def page_context(self):
+        return {
+            'pro_bono_form': self.pro_bono_form,
+            'is_submitted': self.is_submitted,
+        }
+
+    @property
+    def page_url(self):
+        return self.url_name
+
+    def post(self, request, *args, **kwargs):
+        if self.pro_bono_form.is_valid():
+            self.pro_bono_form.process_submission(domain=self.requesting_domain)
+            self.is_submitted = True
+        return self.get(request, *args, **kwargs)
+
+
+class ProBonoStaticView(ProBonoMixin, BasePageView):
+    template_name = 'domain/pro_bono/static.html'
+    urlname = 'pro_bono_static'
+    use_domain_field = True
+
+    @use_select2
+    def dispatch(self, request, *args, **kwargs):
+        return super(ProBonoStaticView, self).dispatch(request, *args, **kwargs)
+
+    @property
+    def requesting_domain(self):
+        return self.pro_bono_form.cleaned_data['domain']
+
+
+class ProBonoView(ProBonoMixin, DomainAccountingSettings):
+    template_name = 'domain/pro_bono/domain.html'
+    urlname = 'pro_bono'
+    use_domain_field = False
+
+    @use_select2
+    def dispatch(self, request, *args, **kwargs):
+        return super(ProBonoView, self).dispatch(request, *args, **kwargs)
+
+    @property
+    def requesting_domain(self):
+        return self.domain
+
+    @property
+    def parent_pages(self):
+        return [
+            {
+                'title': DomainSubscriptionView.page_title,
+                'url': reverse(DomainSubscriptionView.urlname, args=[self.domain]),
+            }
+        ]
+
+    @property
+    def section_url(self):
+        return self.page_url
